# perf: Faster report exporting logic