Try exporting report with 100K rows, most likely it fails or takes a really long time.

Root causes:
- visible_idx was a list, lookups are SLOW AF when lists grow to 100k+. Converted this to set which is one time cost but lookups are O(1) after that. 
- visible_idx check is not required when I am exporting entire report. If length matches then ignore visible IDs completely and just send everything. 



|         | Before | After | Improvement |
| ---     | ---    | ---   | ---         |
| Export 85K rows  | 25.99  | 0.77  | ~33x faster |